refactor test fixutres for verify package
