--- packages/verify/src/__tests__/__fixtures__/bundles/v01/index.ts
@@ -1,4 +1,0 @@
-import dsse from './dsse';
-import signature from './signature';
-
-export default { dsse, signature };
