--- packages/verify/src/__tests__/__fixtures__/bundles/v02/index.ts
@@ -1,32 +1,0 @@
-/*
-Copyright 2023 The Sigstore Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-import crypto from 'crypto';
-import signature from './signature';
-
-export default { signature };
-
-const tlogKey = crypto.createPublicKey(`-----BEGIN PUBLIC KEY-----
-MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE2G2Y+2tabdTV5BcGiBIx0a9fAFwr
-kBbmLSGtks4L3qX6yYY0zufBnhC8Ur/iy55GhWP/9A/bY2LhC30M9+RYtw==
------END PUBLIC KEY-----`);
-
-const tlogKeyID = crypto
-  .createHash('sha256')
-  .update(tlogKey.export({ format: 'der', type: 'spki' }))
-  .digest()
-  .toString('hex');
-
-export const tlogKeys = { [tlogKeyID]: tlogKey };
