+++ packages/verify/src/__tests__/tlog/intoto.test.ts
@@ -13,17 +13,19 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-import { bundleFromJSON } from '@sigstore/bundle';
+import { BundleWithDsseEnvelope, bundleFromJSON } from '@sigstore/bundle';
 import { signatureContent } from '../../bundle';
 import { VerificationError } from '../../error';
 import { verifyIntotoTLogBody } from '../../tlog/intoto';
-import bundles from '../__fixtures__/bundles/v01';
+import * as bundles from '../__fixtures__/bundles';
 
 import type { ProposedIntotoEntry } from '@sigstore/rekor-types';
 
 describe('verifyIntotoTLogBody', () => {
   describe('when everything is valid', () => {
-    const bundle = bundleFromJSON(bundles.dsse.valid.withSigningCert);
+    const bundle = bundleFromJSON(
+      bundles.V1.DSSE.WITH_SIGNING_CERT.TLOG_INTOTO
+    );
     const tlogEntry = bundle.verificationMaterial.tlogEntries[0];
     const body: ProposedIntotoEntry = JSON.parse(
       tlogEntry.canonicalizedBody.toString('utf8')
@@ -36,12 +38,19 @@ describe('verifyIntotoTLogBody', () => {
   });
 
   describe('when the payload hash does NOT match the value in the intoto entry', () => {
-    const bundle = bundleFromJSON(bundles.dsse.invalid.badSignatureTLogIntoto);
+    const bundle = bundleFromJSON(
+      bundles.V1.DSSE.WITH_SIGNING_CERT.TLOG_INTOTO
+    );
     const tlogEntry = bundle.verificationMaterial.tlogEntries[0];
     const body: ProposedIntotoEntry = JSON.parse(
       tlogEntry.canonicalizedBody.toString('utf8')
     );
     const content = signatureContent(bundle);
+
+    beforeEach(() => {
+      (bundle as BundleWithDsseEnvelope).content.dsseEnvelope.payload =
+        Buffer.from('oops');
+    });
 
     it('throws an error', () => {
       expect(() => verifyIntotoTLogBody(body, content)).toThrowWithCode(
@@ -53,13 +62,19 @@ describe('verifyIntotoTLogBody', () => {
 
   describe('when the signature does NOT match the value in the intoto entry', () => {
     const bundle = bundleFromJSON(
-      bundles.dsse.invalid.tlogIntotoIncorrectSigInBody
+      bundles.V1.DSSE.WITH_SIGNING_CERT.TLOG_INTOTO
     );
     const tlogEntry = bundle.verificationMaterial.tlogEntries[0];
     const body: ProposedIntotoEntry = JSON.parse(
       tlogEntry.canonicalizedBody.toString('utf8')
     );
     const content = signatureContent(bundle);
+
+    beforeEach(() => {
+      /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+      (body.spec.content.envelope as any).signatures[0].sig =
+        Buffer.from('oops').toString('base64');
+    });
 
     it('throws an error', () => {
       expect(() => verifyIntotoTLogBody(body, content)).toThrowWithCode(
@@ -70,7 +85,9 @@ describe('verifyIntotoTLogBody', () => {
   });
 
   describe('when the tlog entry version is unsupported', () => {
-    const bundle = bundleFromJSON(bundles.dsse.valid.withSigningCert);
+    const bundle = bundleFromJSON(
+      bundles.V1.DSSE.WITH_SIGNING_CERT.TLOG_INTOTO
+    );
     const tlogEntry = bundle.verificationMaterial.tlogEntries[0];
     const body: ProposedIntotoEntry = JSON.parse(
       tlogEntry.canonicalizedBody.toString('utf8')
@@ -92,13 +109,19 @@ describe('verifyIntotoTLogBody', () => {
 
   describe('when the signature count does NOT match the intoto entry', () => {
     const bundle = bundleFromJSON(
-      bundles.dsse.invalid.tlogIntotoTooManySigsInBody
+      bundles.V1.DSSE.WITH_SIGNING_CERT.TLOG_INTOTO
     );
     const tlogEntry = bundle.verificationMaterial.tlogEntries[0];
     const body: ProposedIntotoEntry = JSON.parse(
       tlogEntry.canonicalizedBody.toString('utf8')
     );
     const content = signatureContent(bundle);
+
+    beforeEach(() => {
+      /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
+      const env = body.spec.content.envelope as any;
+      env.signatures.push(env.signatures[0].sig);
+    });
 
     it('throws an error', () => {
       expect(() => verifyIntotoTLogBody(body, content)).toThrowWithCode(
@@ -109,7 +132,9 @@ describe('verifyIntotoTLogBody', () => {
   });
 
   describe('when the tlog entry body is missing the payload hash', () => {
-    const bundle = bundleFromJSON(bundles.dsse.valid.withSigningCert);
+    const bundle = bundleFromJSON(
+      bundles.V1.DSSE.WITH_SIGNING_CERT.TLOG_INTOTO
+    );
     const tlogEntry = bundle.verificationMaterial.tlogEntries[0];
     const body: ProposedIntotoEntry = JSON.parse(
       tlogEntry.canonicalizedBody.toString('utf8')
